+++ gc.c
@@ -1090,12 +1090,12 @@ rb_data_object_alloc(VALUE klass, void *datap, RUBY_DATA_FUNC dmark, RUBY_DATA_F
 #define STACK_LEVEL_MAX (th->machine_stack_maxsize/sizeof(VALUE))
 
 #if STACK_GROW_DIRECTION < 0
-# define STACK_LENGTH  (STACK_START - STACK_END)
+# define STACK_LENGTH  (size_t)(STACK_START - STACK_END)
 #elif STACK_GROW_DIRECTION > 0
-# define STACK_LENGTH  (STACK_END - STACK_START + 1)
+# define STACK_LENGTH  (size_t)(STACK_END - STACK_START + 1)
 #else
-# define STACK_LENGTH  ((STACK_END < STACK_START) ? STACK_START - STACK_END\
-                                           : STACK_END - STACK_START + 1)
+# define STACK_LENGTH  ((STACK_END < STACK_START) ? (size_t)(STACK_START - STACK_END) \
+			: (size_t)(STACK_END - STACK_START + 1))
 #endif
 #if !STACK_GROW_DIRECTION
 int ruby_stack_grow_direction;
